upd: restructured loggers,  changed some definitions,  and changed go-side call strategy
