+++ pyloggo/__init__.py
@@ -1,16 +1,17 @@
-from .logger import Logger, GlobalLogger
+from .logger.logger import _Logger
+from .logger._global import Logger 
 from .style import FormatStyle
 from .formatter import TextFormatter, JsonFormatter
 from .writer import StdoutWriter, FileWriter, Writer
 from .route import RouteProcessor
 from .enums import LogLevel
 
-logger: GlobalLogger = GlobalLogger()
+logger: Logger = Logger()
 
 
 __all__ = [
     "logger",
-    "Logger",
+    "_Logger",
     "FormatStyle",
     "TextFormatter",
     "JsonFormatter",
